@@ -1,86 +1,106 @@
 {
-  "_args": [
-    [
-      "uuid@3.4.0",
-      "/Users/nickamorim/dev/university/SYDE361/collabeat"
-    ]
-  ],
-  "_from": "uuid@3.4.0",
-  "_id": "uuid@3.4.0",
+  "_from": "uuid",
+  "_id": "uuid@8.2.0",
   "_inBundle": false,
-  "_integrity": "sha512-HjSDRw6gZE5JMggctHBcjVak08+KEVhSIiDzFnT9S9aegmp85S/bReBVTb4QTFaRNptJ9kuYaNhnbNEOkbKb/A==",
+  "_integrity": "sha512-CYpGiFTUrmI6OBMkAdjSDM0k5h8SkkiTP4WAjQgDgNB1S3Ou9VBEvr6q0Kv2H1mMk7IWfxYGpMH5sd5AvcIV2Q==",
   "_location": "/uuid",
   "_phantomChildren": {},
   "_requested": {
-    "type": "version",
+    "type": "tag",
     "registry": true,
-    "raw": "uuid@3.4.0",
+    "raw": "uuid",
     "name": "uuid",
     "escapedName": "uuid",
-    "rawSpec": "3.4.0",
+    "rawSpec": "",
     "saveSpec": null,
-    "fetchSpec": "3.4.0"
+    "fetchSpec": "latest"
   },
   "_requiredBy": [
-    "/request",
-    "/sockjs",
-    "/webpack-log"
+    "#USER",
+    "/"
   ],
-  "_resolved": "https://registry.npmjs.org/uuid/-/uuid-3.4.0.tgz",
-  "_spec": "3.4.0",
-  "_where": "/Users/nickamorim/dev/university/SYDE361/collabeat",
+  "_resolved": "https://registry.npmjs.org/uuid/-/uuid-8.2.0.tgz",
+  "_shasum": "cb10dd6b118e2dada7d0cd9730ba7417c93d920e",
+  "_spec": "uuid",
+  "_where": "/Users/smavehyiashahid/Desktop/SYDE361/collabeat",
   "bin": {
-    "uuid": "bin/uuid"
+    "uuid": "dist/bin/uuid"
   },
   "browser": {
-    "./lib/rng.js": "./lib/rng-browser.js",
-    "./lib/sha1.js": "./lib/sha1-browser.js",
-    "./lib/md5.js": "./lib/md5-browser.js"
+    "./dist/md5.js": "./dist/md5-browser.js",
+    "./dist/rng.js": "./dist/rng-browser.js",
+    "./dist/sha1.js": "./dist/sha1-browser.js",
+    "./dist/esm-node/index.js": "./dist/esm-browser/index.js"
   },
   "bugs": {
     "url": "https://github.com/uuidjs/uuid/issues"
   },
+  "bundleDependencies": false,
   "commitlint": {
     "extends": [
       "@commitlint/config-conventional"
     ]
   },
-  "contributors": [
-    {
-      "name": "Robert Kieffer",
-      "email": "robert@broofa.com"
-    },
-    {
-      "name": "Christoph Tavan",
-      "email": "dev@tavan.de"
-    },
-    {
-      "name": "AJ ONeal",
-      "email": "coolaj86@gmail.com"
-    },
-    {
-      "name": "Vincent Voyer",
-      "email": "vincent@zeroload.net"
-    },
-    {
-      "name": "Roman Shtylman",
-      "email": "shtylman@gmail.com"
-    }
-  ],
+  "deprecated": false,
   "description": "RFC4122 (v1, v4, and v5) UUIDs",
   "devDependencies": {
-    "@commitlint/cli": "~8.2.0",
-    "@commitlint/config-conventional": "~8.2.0",
-    "eslint": "~6.4.0",
-    "husky": "~3.0.5",
-    "mocha": "6.2.0",
-    "runmd": "1.2.1",
-    "standard-version": "7.0.0"
+    "@babel/cli": "7.10.3",
+    "@babel/core": "7.10.3",
+    "@babel/preset-env": "7.10.3",
+    "@commitlint/cli": "9.0.1",
+    "@commitlint/config-conventional": "9.0.1",
+    "@rollup/plugin-node-resolve": "8.0.1",
+    "@wdio/browserstack-service": "6.1.15",
+    "@wdio/cli": "6.1.20",
+    "@wdio/jasmine-framework": "6.1.17",
+    "@wdio/local-runner": "6.1.20",
+    "@wdio/spec-reporter": "6.1.14",
+    "@wdio/static-server-service": "6.1.14",
+    "@wdio/sync": "6.1.14",
+    "babel-eslint": "10.1.0",
+    "bundlewatch": "0.2.7",
+    "eslint": "7.3.0",
+    "eslint-config-prettier": "6.11.0",
+    "eslint-config-standard": "14.1.1",
+    "eslint-plugin-import": "2.21.2",
+    "eslint-plugin-node": "11.1.0",
+    "eslint-plugin-prettier": "3.1.4",
+    "eslint-plugin-promise": "4.2.1",
+    "eslint-plugin-standard": "4.0.1",
+    "husky": "4.2.5",
+    "jest": "25.5.4",
+    "lint-staged": "10.2.11",
+    "npm-run-all": "4.1.5",
+    "prettier": "2.0.5",
+    "rollup": "2.18.0",
+    "rollup-plugin-terser": "6.1.0",
+    "runmd": "1.3.2",
+    "standard-version": "8.0.0"
   },
+  "exports": {
+    ".": {
+      "node": {
+        "module": "./dist/esm-node/index.js",
+        "require": "./dist/index.js",
+        "import": "./wrapper.mjs"
+      },
+      "default": "./dist/esm-browser/index.js"
+    },
+    "./package.json": "./package.json"
+  },
+  "files": [
+    "CHANGELOG.md",
+    "CONTRIBUTING.md",
+    "LICENSE.md",
+    "README.md",
+    "dist",
+    "wrapper.mjs"
+  ],
   "homepage": "https://github.com/uuidjs/uuid#readme",
   "husky": {
     "hooks": {
-      "commit-msg": "commitlint -E HUSKY_GIT_PARAMS"
+      "commit-msg": "commitlint -E HUSKY_GIT_PARAMS",
+      "pre-commit": "lint-staged"
     }
   },
   "keywords": [
